Univariate bias correction via quantile delta mapping

Univariate bias correction based on the quantile delta mapping QDM version of the quantile mapping algorithm from Cannon et al. (2015). QDM constrains model-projected changes in quantiles to be preserved following bias correction by quantile mapping.

MBC_QDM(
  o.c,
  m.c,
  m.p,
  ratio = FALSE,
  trace = 0.05,
  trace.calc = 0.5 * trace,
  jitter.factor = 0,
  n.tau = length(m.p),
  ratio.max = 2,
  ratio.max.trace = 10 * trace,
  ECBC = FALSE,
  ties = "first",
  subsample = NULL,
  pp.type = 7
)

Arguments

o.c

vector of observed samples during the calibration period.

m.c

vector of model outputs during the calibration period.

m.p

vector of model outputs during the projected period.

ratio

logical value indicating if samples are of a ratio quantity (e.g., precipitation).

trace

numeric value indicating the threshold below which values of a ratio quantity (e.g., ratio=TRUE) should be considered exact zeros.

trace.calc

numeric value of a threshold used internally when handling of exact zeros; defaults to one half of trace.

jitter.factor

optional strength of jittering to be applied when quantities are quantized.

n.tau

number of quantiles used in the quantile mapping; NULL equals the length of the m.p series.

ratio.max

numeric value indicating the maximum proportional change allowed for ratio quantities below the ratio.max.trace threshold.

ratio.max.trace

numeric value of a trace threshold used to constrain the proportional change in ratio quantities to ratio.max; defaults to ten times trace.

ECBC

logical value indicating whether mhat.p outputs should be ordered according to o.c ranks, i.e., as in the empirical copula-bias correction (ECBC) algorithm.

ties

method used to handle ties when calculating ordinal ranks.

subsample

use subsample draws of size n.tau to calculate empirical quantiles; if NULL, calculate normally.

pp.type

type of plotting position used in quantile.

Value

a list of with elements consisting of:

mhat.c: vector of bias corrected m.c values for the calibration period.

mhat.p: vector of bias corrected m.p values for the projection period.

References

Cannon, A.J., S.R. Sobie, and T.Q. Murdock, 2015. Bias correction of simulated precipitation by quantile mapping: How well do methods preserve relative changes in quantiles and extremes? Journal of Climate, 28:6938-6959. doi:10.1175/JCLI-D-14-00754.1

See also