formatQM

This function formats the inputs and gets basic statistics for the different Quantile Mapping (QM, DQM, QDM, UQM and SDM) methods available in the climQMBC package. If monthly data is specified, the input series will be reshaped to a matrix of 12 rows and several columns equal to the number of years of each series. If annual data is specified, the input is reshaped to a row vector with same entries as the input series. For precipitation, physically null values (values below pp_threshold) are replaced by random positive values below pp_factor.

formatQM(obs, mod, var, frq = "M", pp_threshold = 1, pp_factor = 0.01)

Arguments

obs, mod

A column vector of monthly or annual observed (modeled) data (temperature or precipitation).

If monthly frequency is specified, the length of this vector is 12 times the number of years [12 x ny].

If annual frequency is specified, the length of this vector is equal to the number of years [ny].

var

A flag that identifies if data are temperature or precipitation. This flag tells the getDist function if it has to discard distribution functions that allow negative numbers, and if the terms in the correction equations are multiplied/divided or added/subtracted.

Temperature : var = 0;

Precipitation: var = 1.

frq

A string specifying if the input is annual or monthly data. If not specified, it is set monthly as default. Monthly: frq = 'M'; Annual: frq = 'A'

pp_threshold

A float indicating the threshold to consider physically null precipitation values.

pp_factor

A float indicating the maximum value of the random values that replace physically null precipitation values.

Value

list(obs, mod), obs and mod have the same structure:

data:

A column vector of monthly or annual observed data (temperature or precipitation). If monthly frequency is specified, the length of this vector is 12 times the number of observed years [12, y_obs]. If annual frequency is specified, the length of this vector is equal to the number of observed years [1, y_obs].

coef:

mu: A column vector of mean values of the series. [12,1] if the series consider monthly data and [1,1] if the series consider annual data.

sigma: A column vector of standard deviation of the series. [12,1] if the series consider monthly data and [1,1] if the series consider annual data.

skew: A column vector of skewness of the series. [12,1] if the series consider monthly data and [1,1] if the series consider annual data.

skewy: A column vector of skewness of the logarithm of the series. [12,1] if the series consider monthly data and [1,1] if the series consider annual data.

nyear

length of year