Get probability distribution function

Get probability distribution function for each month of the period.

For monthly data, it will have 12 rows and each row will represent a month.

For annual data the series will have only one row.

This function assigns an independent probability distribution function to each row of the input series by comparing the empirical probability distribution function with seven distributions based on the Kolmogorov-Smirnov (KS) test.

Only strictly positive distributions are considered for precipitation and strictly positive distributions are discarded if the series has negative values.

getDist(series, var, coef)

Arguments

series

A matrix of monthly or annual data (temperature or precipitation). If the series consider monthly data, it will have 12 rows and each row will represent a month. For annual data the series will have only one row.

var

A flag that identifies if data are temperature or precipitation. This flag tells the getDist function if it has to discard distribution functions that allow negative numbers.

Temperature: var = 0;

Precipitation: var = 1, which not allow negative numbers.

coef

A named list, with the elements of:

mu: A column vector of mean values of the series. [12,1] if the series consider monthly data and [1,1] if the series consider annual data.

sigma: A column vector of standard deviation of the series. [12,1] if the series consider monthly data and [1,1] if the series consider annual data.

skew: A column vector of skewness of the series. [12,1] if the series consider monthly data and [1,1] if the series consider annual data.

skewy: A column vector of skewness of the logarithm of the series. [12,1] if the series consider monthly data and [1,1] if the series consider annual data.

Value

PDF: A column vector with an ID for the resulting distribution from the KS test. [12,1] if the series consider monthly data and [1,1] if the series consider annual data.

The ID is related to the numeration of the distribution listed in the description of this function. This ID is used in the getCDF and getCDFinv

Distributions

The available distributions are:

Normal distribution;

Log-Normal distribution;

Gamma 2 parameters distribution;

Gamma 3 parameters distribution (Pearson 3 parameters distribution);

Log-Gamma 3 parameters distribution (Log-Pearson 3 parameters distribution);

Gumbel distribution;

Exponential distribution

For precipitation, only 2), 3) and 5) are considered (1, 4, 6, and 7 are discarded).

For series with negative values, only 1), 3), 4), 6), and 7) are considered (2, 3 and 5 are discarded).

Examples

# temperature data
var = 0
input <- clim_Temp

l <- with(input, formatQM(obs, mod, var))
print(str(l))
#> List of 2
#>  $ obs:List of 3
#>   ..$ data : num [1:12, 1:36] 20.5 19.5 17.6 14.5 11.7 ...
#>   ..$ coef :List of 4
#>   .. ..$ mu   : num [1:12, 1] 20.4 19.9 18.1 14.4 11.2 ...
#>   .. ..$ sigma: num [1:12, 1] 0.594 0.614 0.689 0.653 0.767 ...
#>   .. ..$ skew : num [1:12, 1] -0.0181 -0.0193 0.3984 0.769 0.0744 ...
#>   .. ..$ skewy: num [1:12, 1] -0.0722 -0.0838 0.3203 0.5709 -0.0809 ...
#>   ..$ nyear: num 36
#>  $ mod:List of 3
#>   ..$ data : num [1:12, 1:122] 18.8 20.1 19.7 15 11 ...
#>   ..$ coef :List of 4
#>   .. ..$ mu   : num [1:12, 1] 21.1 21.3 19 15.1 11.6 ...
#>   .. ..$ sigma: num [1:12, 1] 0.987 1.1 1.272 1.085 0.876 ...
#>   .. ..$ skew : num [1:12, 1] -0.4576 -0.0376 0.2044 0.4213 -0.2204 ...
#>   .. ..$ skewy: num [1:12, 1] -0.566 -0.196 0.113 0.21 -0.318 ...
#>   ..$ nyear: num 122
#> NULL

l_obs <- l$obs
l_mod <- l$mod

DIST_mod <- with(l_mod, getDist(data, var, coef))

cdf_mod <- with(l_mod, getCDF(DIST_mod, data, coef))

DIST_obs <- with(l_obs, getDist(data, var, coef))
QM_series <- getCDFinv(DIST_obs, cdf_mod, l_obs$coef)

r <- list(dist = DIST_mod, cdf = cdf_mod)
print(str(r))
#> List of 2
#>  $ dist: num [1:12, 1] 2 3 1 1 1 1 1 5 5 4 ...
#>  $ cdf : num [1:12, 1:122] 0.00835 0.14668 0.7073 0.4453 0.26057 ...
#> NULL